requirements.txt
pandas
numpy
streamlit
openai
python-dotenv
rapidfuzz
